% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rcolombos.R
\name{advanced_search_by_both}
\alias{advanced_search_by_both}
\title{Accessory function allowing the advanced_search by both g_ids and c_ids}
\usage{
advanced_search_by_both(organism, g_ids, geneNames, c_ids, g_search_type,
  ann_type, c_search_type)
}
\arguments{
\item{organism}{A character containing the organism id: use \code{\link{listOrganisms}} to display
the available organisms.}

\item{g_ids}{A vector of strings representing contrast_id, go terms, experiment id or condition id according the search type.}

\item{geneNames}{boolean if == FALSE (default) return the locustag otherwise the gene_name for the selected genes.}

\item{c_ids}{A vector of strings representing contrast_id, go terms, experiment id or condition id according the search type.}

\item{g_search_type}{A string either genes, go or annotation.}

\item{ann_type}{A string containing the selected gene_annotation_type: use \code{\link{listEntities}} to display the available entities.}

\item{c_search_type}{A string either contrast_names. experiment, go, condition.}
}
\value{
A data.frame containing locustag (gene_names),
contrasts and M-values for the current organism and genes.
}
\description{
Accessory function allowing the advanced_search by both g_ids and c_ids
}
\references{
http://colombos.net
}

