\name{listOrganisms}
\alias{listOrganisms}
\title{Returns a character vector corresponding to the currently available organisms.}
\usage{
  listOrganisms()
}
\value{
  A list containing the currently available organisms.
}
\description{
  Returns a character vector corresponding to the currently
  available organisms.
}
\examples{
\dontrun{
library('Rcolombos')
listOrganisms()
}
}
\references{
  http://colombos.net
}

