% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{parseCompendium}
\alias{parseCompendium}
\title{This method allows importing the full compendium for the selected organism from a local file}
\usage{
parseCompendium(destfile)
}
\arguments{
\item{destfile}{A character containing the full path of the downloaded file}
}
\value{
A list containing two or three data.frames.

In case \code{\link{switchVersion}} is equal to 2:
\item{exprdata}{the full compendium for the selected organism}
\item{condannot}{The condition annotation for the selected organism}

In case \code{\link{switchVersion}} is equal to 3:
\item{exprdata}{the full compendium for the selected organism}
\item{refannot}{The condition annotation for the reference contrasts}
\item{testannot}{The condition annotation for the test contrasts}
}
\description{
This method allows importing the full compendium for the selected organism from a local file
}
\examples{
\dontrun{
library('Rcolombos')
mtube <- parseCompendium("mtube_compendium_data.zip")
}
}
\references{
http://colombos.net
}

