% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rcolombos.R
\docType{package}
\name{Rcolombos}
\alias{Rcolombos}
\alias{Rcolombos-package}
\title{Interface to Colombos Compendia using the Exposed REST API}
\description{
Provides programmatic access to Colombos, a web based
interface for exploring and analyzing comprehensive organism-specific
cross-platform expression compendia of bacterial organisms.
}

