% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{switchVersion}
\alias{switchVersion}
\title{Select the COLOMBOS REST API version to be used for retrieving data}
\usage{
switchVersion(version = 3)
}
\arguments{
\item{version}{positive number 2 or 3 - 3 (current REST API version) as default}
}
\description{
Select the COLOMBOS REST API version to be used for retrieving data
}
\examples{
\dontrun{
library('Rcolombos')
switchVersion (version = 2) # switch from COLOMBOS REST API 3 to 2
}
}
\references{
http://colombos.net
}

