% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{listGenes}
\alias{listGenes}
\title{This method takes as parameter a single string, representing an organism,
and returns a character vector corresponding to the currently available organisms.}
\usage{
listGenes(organism = "ecoli")
}
\arguments{
\item{organism}{A character containing the organism id: use \code{\link{listOrganisms}} to display
the available organisms.}
}
\value{
A data.frame containing the locustag and description
of all the genes for the selected organism.
}
\description{
This method takes as parameter a single string, representing an organism,
and returns a character vector corresponding to the currently available organisms.
}
\examples{
\dontrun{
library('Rcolombos')
listGenes()
}
}
\references{
http://colombos.net
}

