% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{listOrganisms}
\alias{listOrganisms}
\title{Returns a character vector corresponding to the currently available organisms.}
\usage{
listOrganisms()
}
\value{
A list containing the currently available organisms.
}
\description{
Returns a character vector corresponding to the currently available organisms.
}
\examples{
\dontrun{
library('Rcolombos')
listOrganisms()
}
}
\references{
http://colombos.net
}

